modulith(phase-6): extract submission module
- Move Submission entity, SubmissionRepository, SubmissionService, persistence adapters, controller, DTOs to submission package
- Decouple Submission JPA: @ManyToOne Team -> @Column Long teamId, @ManyToOne Question -> @Column Long questionId
- Change Submission.grade() to accept (correctAnswer, questionPoints) parameters
- Create SubmissionFacade, SubmissionInfoDto
- Create AnswerSubmittedEvent, SubmissionGradedEvent
- Create SubmissionEventListener for QuestionDeletedEvent + TeamRemovedEvent cascade cleanup
- Add questionExists() to QuizFacade
- Update SubmissionRepository: findByTeamId, findByQuestionId, findByTeamIdAndQuestionId, deleteByQuestionId, deleteByTeamId
- Update SubmissionService to use QuizFacade + TeamFacade instead of direct repository access
- Update GameFlowService and AnswerDistributionService for decoupled Submission
- Update all external references and test files
- 215 tests passing

+++ backend/src/test/java/com/intelliquiz/api/application/services/EntityValidationPropertyTest.java
@@ -3,6 +3,7 @@ package com.intelliquiz.api.application.services;
 import com.intelliquiz.api.domain.entities.*;
 import com.intelliquiz.api.quiz.internal.domain.entities.Quiz;
 import com.intelliquiz.api.quiz.internal.domain.entities.Question;
+import com.intelliquiz.api.submission.internal.domain.entities.Submission;
 import com.intelliquiz.api.team.internal.domain.entities.Team;
 import com.intelliquiz.api.user.internal.domain.entities.User;
 import com.intelliquiz.api.shared.enums.*;
@@ -105,7 +106,7 @@ public class EntityValidationPropertyTest {
      */
     @Property(tries = 20)
     void submissionTimestampCannotBeInFuture(@ForAll("futureTimestamps") LocalDateTime futureTime) {
-        Submission submission = new Submission(null, null, "answer");
+        Submission submission = new Submission((Long) null, (Long) null, "answer");
         submission.setSubmittedAt(futureTime);
         
         assertThatThrownBy(submission::validateSubmittedAt)
@@ -118,7 +119,7 @@ public class EntityValidationPropertyTest {
      */
     @Property(tries = 20)
     void submissionTimestampValidationAcceptsPastOrPresent(@ForAll("pastTimestamps") LocalDateTime pastTime) {
-        Submission submission = new Submission(null, null, "answer");
+        Submission submission = new Submission((Long) null, (Long) null, "answer");
         submission.setSubmittedAt(pastTime);
         
         // Should not throw
